# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * FindFit API – busca visual de moda por foto ou texto
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SearchInput {
  /** @nullable */
  queryText?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
}

export interface Search {
  id: number;
  /** @nullable */
  queryText?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  extractedAttributes?: string | null;
  createdAt: string;
}

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  material?: string | null;
  imageUrl: string;
  price: number;
  priceFormatted?: string;
  storeName: string;
  /** @nullable */
  storeType?: string | null;
  /** @nullable */
  storeUrl?: string | null;
  /** @nullable */
  similarity?: number | null;
  sponsored?: boolean;
  /** @nullable */
  badgeLabel?: string | null;
  isFavorited?: boolean;
}

export interface SearchResultList {
  searchId: number;
  total: number;
  products: Product[];
}

export interface ProductDetail {
  id: number;
  name: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  material?: string | null;
  /** @nullable */
  composition?: string | null;
  /** @nullable */
  shipping?: string | null;
  imageUrl: string;
  price: number;
  priceFormatted?: string;
  storeName: string;
  /** @nullable */
  storeCollection?: string | null;
  /** @nullable */
  storeType?: string | null;
  /** @nullable */
  storeUrl?: string | null;
  sponsored?: boolean;
  isFavorited?: boolean;
  similar: Product[];
}

export interface Store {
  id: number;
  name: string;
  /** @nullable */
  type?: string | null;
  /** @nullable */
  url?: string | null;
  address: string;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  /** @nullable */
  distanceKm?: number | null;
  productCount: number;
}

export interface Favorite {
  id: number;
  productId: number;
  product?: Product;
  createdAt: string;
}

export interface FavoriteInput {
  productId: number;
}

export interface Category {
  id: number;
  label: string;
  value: string;
}