# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * FindFit API – busca visual de moda por foto ou texto
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  material?: string | null;
  imageUrl: string;
  price: number;
  priceFormatted?: string;
  storeName: string;
  /** @nullable */
  storeType?: string | null;
  /** @nullable */
  storeUrl?: string | null;
  /** @nullable */
  similarity?: number | null;
  sponsored?: boolean;
  /** @nullable */
  badgeLabel?: string | null;
  isFavorited?: boolean;
}