# lib/api-zod/src/generated/types/productDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * FindFit API – busca visual de moda por foto ou texto
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from './product';

export interface ProductDetail {
  id: number;
  name: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  material?: string | null;
  /** @nullable */
  composition?: string | null;
  /** @nullable */
  shipping?: string | null;
  imageUrl: string;
  price: number;
  priceFormatted?: string;
  storeName: string;
  /** @nullable */
  storeCollection?: string | null;
  /** @nullable */
  storeType?: string | null;
  /** @nullable */
  storeUrl?: string | null;
  sponsored?: boolean;
  isFavorited?: boolean;
  similar: Product[];
}