# lib/api-zod/src/generated/types/search.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * FindFit API – busca visual de moda por foto ou texto
 * OpenAPI spec version: 0.1.0
 */

export interface Search {
  id: number;
  /** @nullable */
  queryText?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  extractedAttributes?: string | null;
  createdAt: string;
}