# lib/api-zod/src/generated/types/searchResultList.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * FindFit API – busca visual de moda por foto ou texto
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from './product';

export interface SearchResultList {
  searchId: number;
  total: number;
  products: Product[];
}